% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medike.R
\name{medike}
\alias{medike}
\title{Import Medike Nature Preserve Data}
\usage{
medike(apptoken)
}
\arguments{
\item{apptoken}{from Socrata service}
}
\value{
data frame for hydromet station
}
\description{
This function loads the Medike data and provides
some processing, will pass to analysis function
for daily data
}
