% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutale.R
\name{mutale}
\alias{mutale}
\title{Import Mutale Weir Data}
\usage{
mutale()
}
\value{
data frame for hydromet station
}
\description{
This function loads the Mutale data and provides
some processing, will pass to analysis function
for daily data
}
