% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind.R
\name{wind}
\alias{wind}
\title{Average wind direction}
\usage{
wind(t, v, dr = "D")
}
\arguments{
\item{t}{is angle array}

\item{v}{is speed array}

\item{dr}{is ("D")egrees or ("R")adians, default is degrees}
}
\value{
averaged direction over all data
}
\description{
This function takes a wind direction array (1-D)
and averages the direction.  Returns average over
all data provided.  Defaults to data in degrees.
This is needed because if you average 10 degrees
with 350 degrees, the average is zero, but the
result will be 180.  Uses na.rm = TRUE if there
are NAs in the input data array.
}
