% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.R
\name{daily}
\alias{daily}
\title{Convert and export to daily data}
\usage{
daily(x)
}
\arguments{
\item{x}{data frame from LRLhydromet import command}
}
\value{
daily data
}
\description{
This function takes the output from the site
load functions and processed them to daily
values. Maximum of 96 records per day or every
15 minutes.  Output day is in terms of lubridate
date, days since 01 Jan 1970.  Daily precipitation
is judged by day, temperature is based on 08:00 to
08:00 the following morning.
}
